move runMigrated to be in main and work with os.Args
